Use zig package management to handle deps
Also uncomment flecs code to verify it works.

This requires zig master (currently 0.12.0-dev.2710+c3eb592a3) to work.
Or, well, I assume any 0.12 version will work, the point is it won't
work with zig 0.11

The switch to zig master required some slight API changes and the
necessity for `.root_source_file` to be a zig file. Since this example
project does not include any zig code — except the code the builds the
build graph of course — a dummy file was added. george/main.cpp is now
treated as any other file and added to the list of source files for the
executable artifact